+++ client/src/features/Rush/BalanceGame.tsx
@@ -6,10 +6,10 @@ import CTAButton from "@/components/CTAButton";
 import Scroll from "@/components/Scroll";
 import { COOKIE_TOKEN_KEY } from "@/constants/Auth/token.ts";
 import { ASCEND, ASCEND_DESCEND, SCROLL_MOTION } from "@/constants/animation.ts";
-import CardOptions from "@/features/Rush/BalanceGame/CardOptions.tsx";
-import CountDown from "@/features/Rush/BalanceGame/CountDown.tsx";
-import FinalResult from "@/features/Rush/BalanceGame/FinalResult.tsx";
-import SelectedCard from "@/features/Rush/BalanceGame/SelectedCard.tsx";
+import CardOptions from "@/features/Rush/BalanceGameSection/CardOptions.tsx";
+import CountDown from "@/features/Rush/BalanceGameSection/CountDown.tsx";
+import FinalResult from "@/features/Rush/BalanceGameSection/FinalResult.tsx";
+import SelectedCard from "@/features/Rush/BalanceGameSection/SelectedCard.tsx";
 import { useBalanceGameContext } from "@/hooks/useBalanceGameContext.ts";
 import useCountDown from "@/hooks/useCountDown.ts";
 import { SectionKeyProps } from "@/types/sections.ts";
