+++ client/src/features/Rush/BalanceGameSection/FinalResult.tsx
